add remaining files to commit (remove tables concept from answers tidy up the dtos)
