--- src/main/kotlin/uk/gov/justice/digital/assessments/api/EpisodeOasysAnswersDto.kt
@@ -1,8 +1,0 @@
-package uk.gov.justice.digital.assessments.api
-
-data class EpisodeOasysAnswersDto(val episodeAnswers: List<EpisodeOasysAnswerDto> = listOf())
-
-data class EpisodeOasysAnswerDto(
-  val questionCode: String,
-  val answer: String
-)
