+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/QuestionDto.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.assessments.api
 
 import io.swagger.v3.oas.annotations.media.Schema
+import uk.gov.justice.digital.assessments.api.answers.AnswerDto
 import uk.gov.justice.digital.assessments.jpa.entities.refdata.QuestionEntity
 import java.time.LocalDateTime
 import java.util.UUID
@@ -37,7 +38,7 @@ data class QuestionDto(
   @Schema(description = "List of Reference Data Answers")
   val answerDtos: Collection<AnswerDto>,
 
-) {
+  ) {
 
   companion object {
