+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/answers/AssessmentAnswersDto.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.assessments.api
+package uk.gov.justice.digital.assessments.api.answers
 
 import java.util.UUID
