+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/answers/GPDetailsAnswerDto.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.assessments.api
+package uk.gov.justice.digital.assessments.api.answers
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import uk.gov.justice.digital.assessments.restclient.communityapi.PersonalContact
@@ -35,7 +35,7 @@ data class GPDetailsAnswerDto(
   val telephoneNumber: List<String?> = emptyList(),
 ) {
   companion object {
-    fun from(personalContacts: List<PersonalContact>): List<GPDetailsAnswerDto>? {
+    fun from(personalContacts: List<PersonalContact>): List<GPDetailsAnswerDto> {
       return if (personalContacts.isEmpty()) emptyList()
       else personalContacts.map { from(it) }
     }
