+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/assessments/UpdateAssessmentEpisodeDto.kt
@@ -1,6 +1,7 @@
-package uk.gov.justice.digital.assessments.api
+package uk.gov.justice.digital.assessments.api.assessments
 
 import io.swagger.v3.oas.annotations.media.Schema
+import uk.gov.justice.digital.assessments.api.AnswersDto
 
 class UpdateAssessmentEpisodeDto(
   @Schema(description = "Answers associated with this episode")
