+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/groups/GroupContentDto.kt
@@ -1,8 +1,9 @@
-package uk.gov.justice.digital.assessments.api
+package uk.gov.justice.digital.assessments.api.groups
 
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonSubTypes.Type
 import com.fasterxml.jackson.annotation.JsonTypeInfo
+import uk.gov.justice.digital.assessments.api.answers.CheckboxGroupDto
 
 @JsonTypeInfo(
   use = JsonTypeInfo.Id.NAME,
@@ -12,7 +13,6 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo
 @JsonSubTypes(
   Type(value = GroupQuestionDto::class, name = "question"),
   Type(value = GroupWithContentsDto::class, name = "group"),
-  Type(value = TableQuestionDto::class, name = "tableGroup"),
   Type(value = CheckboxGroupDto::class, name = "checkboxGroup")
 )
 interface GroupContentDto
