+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/groups/GroupSummaryDto.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.assessments.api
+package uk.gov.justice.digital.assessments.api.groups
 
 import io.swagger.v3.oas.annotations.media.Schema
 import uk.gov.justice.digital.assessments.jpa.entities.refdata.GroupSummaryEntity
