+++ src/test/kotlin/uk/gov/justice/digital/assessments/api/GroupSummaryDtoTest.kt
@@ -3,6 +3,7 @@ package uk.gov.justice.digital.assessments.api
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import uk.gov.justice.digital.assessments.api.groups.GroupSummaryDto
 import uk.gov.justice.digital.assessments.jpa.entities.refdata.GroupSummaryEntity
 import java.util.UUID
